20220811 - Advanced C# Class and Lab 7
This is basically my initial commit (not changing the real one for now) so it's got all my previous assignments and my first Project attempts. I'll try to push a new commit every day that I work on these to keep a chronology/ get in the habit of working with git.