Daily commit 2022-0817
This is in the morning before I really got underway with much of anything. It occurs to me that this is a dairy of sorts: I won't share my thoughts directly, but a snapshot in time of how I think and what was giving me trouble with this process. Interesting stuff.